requirements.txt
# Core dependencies
click>=8.0.0
pydrive2>=1.15.0
schedule>=1.2.0
cryptography>=41.0.0
requests>=2.28.0
psutil>=5.9.0
watchdog>=3.0.0

# Optional dependencies
flask>=2.3.0           # Web interface
flask-socketio>=5.3.0  # Real-time updates
python-crontab>=2.7.0  # Cron management
pyyaml>=6.0            # YAML config support
rich>=13.0.0           # Beautiful CLI output
typer>=0.9.0           # Modern CLI framework

# Email notifications
smtplib2>=0.2.1
email-validator>=2.0.0

# Cloud providers
google-api-python-client>=2.95.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
dropbox>=11.36.0
azure-storage-blob>=12.17.0

# Compression
lz4>=4.3.0
zstandard>=0.21.0

# Database (for metadata)
sqlalchemy>=2.0.0
alembic>=1.11.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
